+++ starters/blog/src/components/bio.js
@@ -1,3 +1,10 @@
+/**
+ * Bio component that queries for data
+ * with Gatsby's StaticQuery component
+ *
+ * See: https://www.gatsbyjs.org/docs/static-query/
+ */
+
 import React from "react"
 import { StaticQuery, graphql } from "gatsby"
 import Image from "gatsby-image"