+++ www/src/components/button.js
@@ -19,6 +19,7 @@ const Button = ({
   tiny,
   secondary,
   ondark,
+  tracking,
   ...rest
 }) => {
   const Tag = components[tag || `link`]
@@ -41,8 +42,42 @@ const Button = ({
     },
   }
 
+  const trackingOnClick = e => {
+    if (typeof props.onClick === `function`) {
+      props.onClick(e)
+    }
+
+    let redirect = true
+
+    // Slightly modified logic from the gatsby-plugin-google-analytics
+    // But this one should work with `Link` component as well
+    if (
+      e.button !== 0 ||
+      e.altKey ||
+      e.ctrlKey ||
+      e.metaKey ||
+      e.shiftKey ||
+      e.defaultPrevented
+    ) {
+      redirect = false
+    }
+
+    if (props.target && props.target.toLowerCase() !== `_self`) {
+      redirect = false
+    }
+
+    if (tracking && window.ga) {
+      window.ga(`send`, `event`, {
+        eventCategory: `Outbound Link`,
+        eventAction: `click`,
+        eventLabel: `${tracking} - ${props.to || props.href}`,
+        transport: redirect ? `beacon` : ``,
+      })
+    }
+  }
+
   return (
-    <Tag {...props} css={css}>
+    <Tag {...props} onClick={trackingOnClick} css={css}>
       {children}
       {icon && <>{icon}</>}
     </Tag>