+++ www/src/utils/styles.js
@@ -1,5 +1,5 @@
 import hex2rgba from "hex2rgba"
-import { keyframes } from "react-emotion"
+import { keyframes } from "@emotion/core"
 
 import presets, { colors } from "./presets"
 import { rhythm, scale, options } from "./typography"
@@ -62,11 +62,11 @@ export const buttonStyles = {
     },
     ":after": { content: `''`, display: `block` },
     "& svg": { marginLeft: `.2em` },
-    [presets.Tablet]: {
+    [presets.Md]: {
       ...scale(1 / 5),
       padding: `${rhythm(2 / 6)} ${rhythm(3 / 5)}`,
     },
-    [presets.VHd]: { padding: `${rhythm(1 / 2)} ${rhythm(1)}` },
+    [presets.Xxl]: { padding: `${rhythm(1 / 2)} ${rhythm(1)}` },
   },
   secondary: {
     backgroundColor: `transparent`,
@@ -77,20 +77,20 @@ export const buttonStyles = {
     // borderRadius: presets.radiusLg,
     fontSize: scale(1 / 5).fontSize,
     padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
-    [presets.Tablet]: {
+    [presets.Md]: {
       fontSize: scale(2 / 5).fontSize,
       padding: `${rhythm(2 / 4)} ${rhythm(3 / 5)}`,
     },
-    [presets.VHd]: { padding: `${rhythm(1 / 2)} ${rhythm(1)}` },
+    [presets.Xxl]: { padding: `${rhythm(1 / 2)} ${rhythm(1)}` },
   },
   small: {
     fontSize: scale(-1 / 3).fontSize,
     padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
-    [presets.Tablet]: {
+    [presets.Md]: {
       fontSize: scale(-1 / 6).fontSize,
       padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
     },
-    [presets.VHd]: {
+    [presets.Xxl]: {
       fontSize: scale(-1 / 6).fontSize,
       padding: `${rhythm(2 / 5)} ${rhythm(1 / 2)}`,
     },
@@ -98,16 +98,19 @@ export const buttonStyles = {
   tiny: {
     fontSize: scale(-1 / 3).fontSize,
     padding: `${rhythm(1 / 5)} ${rhythm(1 / 3)}`,
-    [presets.Tablet]: {
+    [presets.Md]: {
       fontSize: scale(-1 / 4).fontSize,
       padding: `${rhythm(1 / 5)} ${rhythm(1 / 3)}`,
     },
-    [presets.VHd]: {
+    [presets.Xxl]: {
       fontSize: scale(-1 / 5).fontSize,
       padding: `${rhythm(1 / 5)} ${rhythm(1 / 3)}`,
     },
   },
-  ondark: { border: `1px solid ${colors.ui.light}` },
+  ondark: {
+    border: `1px solid ${colors.ui.light}`,
+    background: colors.gatsbyDark,
+  },
 }
 
 export const svgStyles = {
@@ -134,6 +137,23 @@ export const svgStyles = {
     },
     "& .svg-fill-gradient-accent-white-top": {
       fill: `url(#accent-white-top)`,
+    },
+  },
+}
+
+// This is an exceptionally bad name
+export const linkStyles = {
+  ...scale(-1 / 6),
+  "&&": {
+    alignItems: `center`,
+    border: 0,
+    boxShadow: `none`,
+    color: colors.gray.calm,
+    display: `flex`,
+    fontWeight: `normal`,
+    "&:hover": {
+      background: `none`,
+      color: colors.gatsby,
     },
   },
 }