+++ packages/gatsby-link/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.13](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-link/compare/gatsby-link@2.0.12...gatsby-link@2.0.13) (2019-02-28)
+
+**Note:** Version bump only for package gatsby-link
+
 ## [2.0.12](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-link/compare/gatsby-link@2.0.11...gatsby-link@2.0.12) (2019-02-22)
 
 **Note:** Version bump only for package gatsby-link