+++ packages/gatsby-source-wordpress/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.41](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.40...gatsby-source-wordpress@3.0.41) (2019-02-28)
+
+### Features
+
+- **gatsby-source-wordpress:** Add "path" field to post/page/custom post type entities ([#11775](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/issues/11775)) ([0b39498](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/commit/0b39498))
+
 ## [3.0.40](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wordpress/compare/gatsby-source-wordpress@3.0.39...gatsby-source-wordpress@3.0.40) (2019-02-25)
 
 **Note:** Version bump only for package gatsby-source-wordpress